# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiCatalogResponse": {
        "properties": {
          "apis": {
            "items": {
              "$ref": "#/components/schemas/CatalogApi"
            },
            "title": "Apis",
            "type": "array"
          }
        },
        "title": "ApiCatalogResponse",
        "type": "object"
      },
      "ApiSpecResponse": {
        "description": "Auf Anfrage geladene Spezifikation einer Schnittstelle. truncated\nmeldet, dass das Dokument am Limit abgeschnitten wurde - statt einen\nstillen Teilinhalt auszugeben (Goldene Regel 7).",
        "properties": {
          "api_id": {
            "title": "Api Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "spec_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Type"
          },
          "spec_url": {
            "title": "Spec Url",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "api_id",
          "spec_url",
          "content"
        ],
        "title": "ApiSpecResponse",
        "type": "object"
      },
...
      "CatalogApi": {
        "description": "Eintrag des kuratierten API-Katalogs (geofact_web/data/api_catalog.yaml).\nlayer_template ist - wo vorhanden - ein einbaufertiges Fragment für\ndie layers-Liste der Szenario-YAML, analog CatalogSource.",
        "properties": {
          "auth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "data_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "documentation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentation Url"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/CatalogApiEndpoint"
            },
            "title": "Endpoints",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "layer_template": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Template"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "protocol": {
            "title": "Protocol",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "spec_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Type"
          },
          "spec_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Url"
          },
          "verified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          }
        },
        "required": [
          "id",
          "label",
          "protocol",
          "base_url"
        ],
        "title": "CatalogApi",
        "type": "object"
      },
      "CatalogApiEndpoint": {
        "description": "Ein Endpunkt/Feature-Type einer Schnittstelle. attributes sind die\nin der Praxis relevanten Felder (nicht die vollständige Liste - die\nsteht in der Spezifikation); attribute_notes erklärt einzelne davon,\ndamit ein Mensch den Endpunkt ohne Spec-Studium einordnen kann.",
        "properties": {
          "attribute_notes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attribute Notes",
            "type": "object"
          },
          "attributes": {
            "items": {
              "type": "string"
            },
            "title": "Attributes",
            "type": "array"
          },
          "geometry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "name"
        ],
        "title": "CatalogApiEndpoint",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Fehlerantwort (``HTTPException``): ``detail`` ist die Meldung.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token aus POST /api/auth/login als 'Authorization: Bearer <token>'. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "scheme": "bearer",
        "type": "http"
      },
      "QueryToken": {
        "description": "Dasselbe Token als Query-Parameter - nur für Ereignisstrom, Rasterbild und Download, die der Browser ohne Authorization-Header abruft. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-API zur interaktiven Demonstration des GeoFACT-Frameworks.",
    "title": "GeoFACT Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/apis": {
      "get": {
        "description": "Kuratierte offene Schnittstellen mit Endpunkten, Attributen, Lizenz und Layer-Fragment. Braucht kein Netz.",
        "operationId": "list_apis_api_apis_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Der Katalog des Servers fehlt oder ist nicht lesbar."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "API-Katalog",
        "tags": [
          "apis"
        ]
      }
    },
    "/api/apis/{api_id}/spec": {
      "get": {
        "description": "Lädt GetCapabilities bzw. OpenAPI der Schnittstelle vom Anbieter. Der Text wird bei 400.000 Zeichen abgeschnitten (truncated).",
        "operationId": "get_api_spec_api_apis__api_id__spec_get",
        "parameters": [
          {
            "in": "path",
            "name": "api_id",
            "required": true,
            "schema": {
              "title": "Api Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSpecResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unbekannte Schnittstelle, oder für sie ist keine Spezifikation hinterlegt."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Der Anbieter hat die Spezifikation nicht geliefert."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Spezifikation einer Schnittstelle laden",
        "tags": [
          "apis"
        ]
      }
    },
...
  },
...
}